+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Profiling: fix access to context in case of `NonRecordingSpan`.
+
 ## 1.18.0 - 2024-03-06
 - Upgraded Otel dependencies to 1.23.0 / 0.44b0